+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ 1.2.2 ] - 2024-07-28
+## [ 1.2.2 ] - 2024-08-26
+### Fixed
+- Support of HEAD operation in case of resource delivery.
 
 ## [ 1.2.1 ] - 2024-07-28
 ### Added